add pages for graduateGuide
